refactor(cli): split index.ts — extract lib/repl.ts and lib/welcome.ts
Refs #173. The 433-LOC packages/cli/src/index.ts mixed Commander
program construction, telemetry hooks, REPL, the bare-pax8 welcome
banner, and the argv tokenizer. Pulls REPL and welcome into siblings
under lib/; index.ts now keeps only program construction, hooks,
signal handlers, and the top-level main().

No behavior change. Smoke-checked --version / status --json / error
path with hint to pax8 report-bug all produce byte-identical output.

Final line counts:
- packages/cli/src/index.ts: 229 (was 433)
- packages/cli/src/lib/repl.ts: 143 (new)
- packages/cli/src/lib/welcome.ts: 50 (new)
